# Improve performance of `render_html()`