--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -52,6 +52,7 @@ Imports:
     scales (>= 1.2.1),
     tibble (>= 3.1.8),
     tidyselect (>= 1.2.0),
+    vctrs,
     xml2 (>= 1.3.3)
 Suggests:
     covr,