… NEWS.md
… R/utils_render_rtf.R
… R/z_utils_render_footnotes.R