@rich-iannone 2023-10-18
@mgirlich This is great work and I really appreciate it!

The only problem I can see is that adding `vtcrs` as a dependency would bring the total number to 21, which is something that CRAN flags during updates (and ultimately won't, to my knowledge, accept). However, the `tibble` dependency could be removed with some minor, extra work. If you'd be willing to replace the two instances where tibble is used, then the `vtcrs` addition will be okay.

From a quick search, `tibble` functions are used here:

- https://github.com/rstudio/gt/blob/9c6a46c84c05a111bf9b62e8f07222d2dd79029b/R/utils_render_rtf.R#L415C21-L415C21
- https://github.com/rstudio/gt/blob/9c6a46c84c05a111bf9b62e8f07222d2dd79029b/R/z_utils_render_footnotes.R#L357C5-L357C5